requirements.txt
pytube
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
mtgsdk==1.3.1
numpy==2.3.3
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0

yugioh
sounddevice
osrsbox
scipy